API&Client: Added ingredientDTO, fixed put and add methods, and added logic to bind the measurment type when the item is being updated, and show empty for a new entry.

+++ Client/Models/Ingredient.cs
@@ -56,8 +56,8 @@ namespace Client.Models
 			}
 		}
 
-		int _quantity;
-		public int Quantity
+		float _quantity;
+		public float Quantity
 		{
 			get => _quantity;
 			set
